requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
altair==4.0.1             # via streamlit
appdirs==1.4.3            # via black
appnope==0.1.0            # via ipykernel, ipython
astor==0.8.1              # via streamlit
attrs==19.3.0             # via black, jsonschema, pandas-vet
backcall==0.1.0           # via ipython
base58==2.0.0             # via streamlit
black==19.10b0            # via -r requirements.in
bleach==3.1.3             # via nbconvert
blinker==1.4              # via streamlit
boto3==1.12.28            # via streamlit
botocore==1.15.28         # via boto3, s3transfer, streamlit
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via black, streamlit
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via ipython, traitlets, validators
defusedxml==0.6.0         # via nbconvert
docutils==0.15.2          # via botocore, python-daemon
entrypoints==0.3          # via altair, flake8, nbconvert
enum-compat==0.0.3        # via streamlit
flake8==3.7.9             # via -r requirements.in, pandas-vet
future==0.18.2            # via streamlit
idna==2.9                 # via requests
importlib-metadata==1.5.1  # via jsonschema
ipykernel==5.2.0          # via ipywidgets, jupyter, jupyter-console, notebook, pydeck, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==7.13.0           # via ipykernel, ipywidgets, jupyter-console
ipywidgets==7.5.1         # via jupyter, pydeck
isort==4.3.21             # via -r requirements.in
jedi==0.16.0              # via ipython
jinja2==2.11.1            # via altair, nbconvert, notebook, pydeck
jmespath==0.9.5           # via boto3, botocore
joblib==0.14.1            # via scikit-learn
jsonschema==3.2.0         # via altair, nbformat
jupyter-client==6.1.0     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==6.1.0    # via jupyter
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0            # via -r requirements.in
kiwisolver==1.1.0         # via matplotlib
lockfile==0.12.2          # via python-daemon
luigi==2.8.12             # via -r requirements.in
markupsafe==1.1.1         # via jinja2
matplotlib==3.2.1         # via -r requirements.in, seaborn
mccabe==0.6.1             # via flake8
mistune==0.8.4            # via nbconvert
mypy-extensions==0.4.3    # via mypy
mypy==0.770               # via -r requirements.in
natsort==7.0.1            # via pyjanitor
nbconvert==5.6.1          # via jupyter, notebook
nbformat==5.0.4           # via ipywidgets, nbconvert, notebook
notebook==6.0.3           # via jupyter, widgetsnbextension
numpy==1.18.2             # via -r requirements.in, altair, matplotlib, pandas, pyjanitor, scikit-learn, scipy, seaborn, streamlit, xarray
pandas-flavor==0.2.0      # via pyjanitor
pandas-vet==0.2.2         # via -r requirements.in
pandas==1.0.3             # via -r requirements.in, altair, pandas-flavor, pyjanitor, seaborn, streamlit, xarray
pandocfilters==1.4.2      # via nbconvert
parso==0.6.2              # via jedi
pathspec==0.7.0           # via black
pathtools==0.1.2          # via watchdog
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.0.0             # via streamlit
prometheus-client==0.7.1  # via notebook
prompt-toolkit==3.0.4     # via ipython, jupyter-console
protobuf==3.11.3          # via streamlit
ptyprocess==0.6.0         # via pexpect, terminado
pycodestyle==2.5.0        # via flake8
pydeck==0.2.1             # via streamlit
pyflakes==2.1.1           # via flake8
pygments==2.6.1           # via ipython, jupyter-console, nbconvert, qtconsole
pyjanitor==0.20.5         # via -r requirements.in
pyparsing==2.4.6          # via matplotlib
pyrsistent==0.16.0        # via jsonschema
python-daemon==2.2.4      # via luigi
python-dateutil==2.8.0    # via botocore, jupyter-client, luigi, matplotlib, pandas, streamlit
pytz==2019.3              # via pandas, tzlocal
pyzmq==19.0.0             # via jupyter-client, notebook, qtconsole
qtconsole==4.7.2          # via jupyter
qtpy==1.9.0               # via qtconsole
regex==2020.2.20          # via black
requests==2.23.0          # via pyjanitor, streamlit
s3transfer==0.3.3         # via boto3
scikit-learn==0.22.2.post1  # via pyjanitor
scipy==1.4.1              # via -r requirements.in, scikit-learn, seaborn
seaborn==0.10.0           # via -r requirements.in
send2trash==1.5.0         # via notebook
six==1.14.0               # via bleach, cycler, jsonschema, protobuf, pyrsistent, python-dateutil, traitlets, validators
streamlit==0.56.0         # via -r requirements.in
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
toml==0.10.0              # via black, streamlit
toolz==0.10.0             # via altair
tornado==5.1.1            # via ipykernel, jupyter-client, luigi, notebook, streamlit, terminado
traitlets==4.3.3          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, pydeck, qtconsole
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.1  # via mypy
tzlocal==2.0.0            # via streamlit
urllib3==1.25.8           # via botocore, requests
validators==0.14.2        # via streamlit
watchdog==0.10.2          # via streamlit
wcwidth==0.1.9            # via prompt-toolkit
webencodings==0.5.1       # via bleach
widgetsnbextension==3.5.1  # via ipywidgets
xarray==0.15.1            # via pandas-flavor
xlrd==1.2.0               # via pyjanitor
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
